feat: refactor code, add doc comments, improve error handling
