+++ backend/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "backend"
+name = "scottylabs_bus_backend"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
 axum = "0.7"
-reqwest = { version = "0.11", features = ["json"] }
+reqwest = { version = "0.12", features = ["json"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 tokio = { version = "1", features = ["full"] }
